+++ assets/js/libs/htmx-extensions/ajax-header/ajax-header.js
@@ -1,9 +1,0 @@
-(function() {
-  htmx.defineExtension('ajax-header', {
-    onEvent: function(name, evt) {
-      if (name === 'htmx:configRequest') {
-        evt.detail.headers['X-Requested-With'] = 'XMLHttpRequest'
-      }
-    }
-  })
-})()